Move printing errors to where we have more player info
Also just filter out players that fail to parse.

I'd be good to scan the output and make sure none jump out
(they're probably special cases I haven't caught) but the list
from DraftKings is much longer than the list of players on
FantasyPros. This results in a bunch of "failures" that are
actually 4th string TEs that aren't relevant anyways.